+++ packages/elysia/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/elysia
 
+## 0.1.101
+
+### Patch Changes
+
+- @vercel/node@5.8.25
+
 ## 0.1.100
 
 ### Patch Changes