+++ packages/gatsby-plugin-vercel-builder/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vercel/gatsby-plugin-vercel-builder
 
+## 2.2.27
+
+### Patch Changes
+
+- Updated dependencies [def07fc]
+  - @vercel/build-utils@13.33.1
+
 ## 2.2.26
 
 ### Patch Changes