+++ packages/hono/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/hono
 
+## 0.2.104
+
+### Patch Changes
+
+- @vercel/node@5.8.25
+
 ## 0.2.103
 
 ### Patch Changes